Merging mem_tracker and draw_state layers into core_validation layer
Currently it compiles for Linux and Android and cube/tri --validate are working

Very little efficiency has been pulled out for this initial merge.
A few maps have been merged, but otherwise almost all of the mem_tracker
code is pulled intact and surrounded in "#if MTMERGE" blocks.

core_validation.cpp is up-to-date w/ master as of 3/15 at 9am MST.